rag_chatbot/requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
faiss-cpu
pypdf
google-generativeai
